+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 * Add `repl.fork` to start run and use an Anvil instance as a fork of the current URL
 * Add `repl.startPrank` / `repl.stopPrank` to start/stop impersonating an address
-* Add `FUNC.trace_call` method to contract functions
+* Add `FUNC.traceCall` method to contract functions
 * Add `abi.decodeData` to decode function calldata and errors from any known ABI
 
 ### Other changes